+++ rasa_core/server.py
@@ -18,6 +18,8 @@ from typing import Union, Text, Optional
 
 from rasa_core import utils, events
 from rasa_core.agent import Agent
+from rasa_core.channels.direct import CollectingOutputChannel
+from rasa_core.channels import UserMessage
 from rasa_core.interpreter import NaturalLanguageInterpreter
 from rasa_core.tracker_store import TrackerStore
 from rasa_core.trackers import DialogueStateTracker
@@ -102,6 +104,23 @@ def bool_arg(request, name, default=True):
 
     d = str(default)
     return request.args.get(name, d).lower() == 'true'
+
+
+def request_parameters(request):
+    if request.method.decode('utf-8', 'strict') == 'GET':
+        return {
+            key.decode('utf-8', 'strict'): value[0].decode('utf-8',
+                                                           'strict')
+            for key, value in request.args.items()}
+    else:
+        content = request.content.read()
+        try:
+            return json.loads(content.decode('utf-8', 'strict'))
+        except ValueError as e:
+            logger.error("Failed to decode json during respond request. "
+                         "Error: {}. Request content: "
+                         "'{}'".format(e, content))
+            raise
 
 
 class RasaCoreServer(object):
@@ -255,20 +274,7 @@ class RasaCoreServer(object):
     @ensure_loaded_agent
     def parse(self, request, sender_id):
         request.setHeader('Content-Type', 'application/json')
-        if request.method.decode('utf-8', 'strict') == 'GET':
-            request_params = {
-                key.decode('utf-8', 'strict'): value[0].decode('utf-8',
-                                                               'strict')
-                for key, value in request.args.items()}
-        else:
-            content = request.content.read()
-            try:
-                request_params = json.loads(content.decode('utf-8', 'strict'))
-            except ValueError as e:
-                logger.error("Failed to decode json during parse request. "
-                             "Error: {}. Request content: "
-                             "'{}'".format(e, content))
-                raise
+        request_params = request_parameters(request)
 
         if 'query' in request_params:
             message = request_params.pop('query')
@@ -286,6 +292,37 @@ class RasaCoreServer(object):
             request.setResponseCode(500)
             logger.error("Caught an exception during "
                          "parse: {}".format(e), exc_info=1)
+            return json.dumps({"error": "{}".format(e)})
+
+    @app.route("/conversations/<sender_id>/respond",
+               methods=['GET', 'POST', 'OPTIONS'])
+    @check_cors
+    @requires_auth
+    @ensure_loaded_agent
+    def respond(self, request, sender_id):
+        request.setHeader('Content-Type', 'application/json')
+        request_params = request_parameters(request)
+
+        if 'query' in request_params:
+            message = request_params.pop('query')
+        elif 'q' in request_params:
+            message = request_params.pop('q')
+        else:
+            request.setResponseCode(400)
+            return json.dumps({"error": "Invalid respond parameter specified"})
+
+        try:
+            out = CollectingOutputChannel()
+            responses = self.agent.handle_message(message,
+                                                  output_channel=out,
+                                                  sender_id=sender_id)
+            request.setResponseCode(200)
+            return json.dumps(responses)
+
+        except Exception as e:
+            request.setResponseCode(500)
+            logger.error("Caught an exception during "
+                         "respond: {}".format(e), exc_info=1)
             return json.dumps({"error": "{}".format(e)})
 
     @app.route("/load", methods=['POST', 'OPTIONS'])