+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - supported loading training data from a folder - loads all stories from
   all files in that directory
 - parameter to specify NLU project when instantiating a ``RasaNLUInterpreter``
+- simple ``/respond`` endpoint to get bot response to a user message
 
 Changed
 -------